requirements.txt
pytest-django
django-filter
psycopg2
Pillow
flake8